Modularize code in different files
